Add README
I'm going to put my notes in the README to detail some of the decisions
I've made in the project so that it'll be easier for reviewers to
understand what I'm trying to do, and can give good comments. Also might
be helpful for beginners to read about; but honestly, this isn't really
meant to be a guide to build anything. Take it with a grain of salt.