Add `Index Person` network API call
This took a lot of exploring through `req`'s docs. I feel like it still
could be improved with more examples, but I somehow managed. It felt like
I was digging around the Sahara Desert trying to find 32 skull keys that
together unlock a mysterious cavern that takes me to the next stage.

Or something like that.